🚚 api: move repo to src
