+++ api/src/certification/session/infrastructure/repositories/session-for-supervisor-kit-repository.js
@@ -1,5 +1,5 @@
-import { knex } from '../../../../db/knex-database-connection.js';
-import { SessionForSupervisorKit } from '../../../domain/read-models/SessionForSupervisorKit.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { SessionForSupervisorKit } from '../../domain/read-models/SessionForSupervisorKit.js';
 
 const get = async function (idSession) {
   const results = await knex
