+++ api/src/certification/shared/domain/usecases/index.js
@@ -2,6 +2,7 @@ import { fileURLToPath } from 'node:url';
 import { dirname, join } from 'node:path';
 
 import * as assessmentRepository from '../../../../shared/infrastructure/repositories/assessment-repository.js';
+import * as attendanceSheetPdfUtils from '../../../session/infrastructure/utils/pdf/attendance-sheet-pdf.js';
 import * as badgeRepository from '../../../../../lib/infrastructure/repositories/badge-repository.js';
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
 import * as certificationChallengeLiveAlertRepository from '../../../session/infrastructure/repositories/certification-challenge-live-alert-repository.js';
@@ -11,16 +12,16 @@ import * as complementaryCertificationBadgesRepository from '../../../complement
 import * as complementaryCertificationRepository from '../../../../../lib/infrastructure/repositories/complementary-certification-repository.js';
 import * as complementaryCertificationForTargetProfileAttachmentRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-for-target-profile-attachment-repository.js';
 import * as complementaryCertificationTargetProfileHistoryRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-target-profile-history-repository.js';
+import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
 import * as issueReportCategoryRepository from '../../../shared/infrastructure/repositories/issue-report-category-repository.js';
 import * as mailService from '../services/mail-service.js';
 import * as organizationRepository from '../../../complementary-certification/infrastructure/repositories/organization-repository.js';
 import * as sessionCodeService from '../../../session/domain/services/session-code-service.js';
+import * as sessionForAttendanceSheetRepository from '../../../session/infrastructure/repositories/session-for-attendance-sheet-repository.js';
+import * as sessionForSupervisorKitRepository from '../../../session/infrastructure/repositories/session-for-supervisor-kit-repository.js';
+import * as sessionRepository from '../../../session/infrastructure/repositories/session-repository.js';
 import * as sessionValidator from '../../../session/domain/validators/session-validator.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
-import * as sessionRepository from '../../../session/infrastructure/repositories/session-repository.js';
-import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
-import * as sessionForAttendanceSheetRepository from '../../../session/infrastructure/repositories/session-for-attendance-sheet-repository.js';
-import * as attendanceSheetPdfUtils from '../../../session/infrastructure/utils/pdf/attendance-sheet-pdf.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 
@@ -38,11 +39,12 @@ const dependencies = {
   complementaryCertificationTargetProfileHistoryRepository,
   flashAlgorithmService,
   issueReportCategoryRepository,
+  mailService,
   organizationRepository,
   sessionCodeService,
-  mailService,
-  sessionRepository,
   sessionForAttendanceSheetRepository,
+  sessionForSupervisorKitRepository,
+  sessionRepository,
   sessionValidator,
   userRepository,
 };
